Reorganized code, added team names from statbotics :D
