+++ lib/pages/summation/averages/averages_page.dart
@@ -3,20 +3,17 @@ import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:scouting_site/pages/summation/averages/comaprasion_page.dart';
 
 // Project imports:
+import 'package:scouting_site/pages/summation/averages/comaprasion_page.dart';
+import 'package:scouting_site/pages/summation/averages/insights_page.dart';
 import 'package:scouting_site/pages/summation/scouting_entries_page.dart';
-import 'package:scouting_site/pages/summation/team_overview.dart';
 import 'package:scouting_site/services/firebase/firebase_api.dart';
 import 'package:scouting_site/services/scouting/form_data.dart';
 import 'package:scouting_site/services/scouting/form_page_data.dart';
-import 'package:scouting_site/services/scouting/helper_methods.dart';
 import 'package:scouting_site/services/scouting/question.dart';
 import 'package:scouting_site/services/scouting/scouting.dart';
 import 'package:scouting_site/theme.dart';
-import 'package:scouting_site/widgets/dialog_widgets/dialog_text_figures.dart';
-import 'package:scouting_site/widgets/dialog_widgets/dialog_text_input.dart';
 
 class AveragesPage extends StatefulWidget {
   List<FormData>? formsData;
@@ -31,60 +28,10 @@ class _AveragesPageState extends State<AveragesPage> {
   List<FormData> _formsData = [];
   Map<String, double> _pageAvgs = {};
   double _totalAllTeamsAvg = 0;
-  String _sortBy = "total_score";
-  Map<String, dynamic> _searchQueryMap = {};
 
   @override
   Widget build(BuildContext context) {
-    if (widget.formsData == null) {
-      getDocuments();
-    } else {
-      _formsData = widget.formsData!;
-    }
-
-    _formsData = _formsData.map((match) {
-      if (match.matchType != MatchType.normal && (match.game ?? 0) < 200) {
-        int addition =
-            MatchType.rematch == (match.game ?? MatchType.normal) ? 200 : 300;
-        return match..game = (match.game ?? 0) + addition;
-      } else {
-        return match;
-      }
-    }).toList();
-
-    _formsData = calculateAvgs(_formsData).values.toList();
-
-    switch (_sortBy) {
-      case "total_score":
-        sortByTotalScore(_formsData);
-        break;
-      case "game":
-        sortByGame(_formsData);
-        break;
-      case "team":
-        sortByTeam(_formsData);
-        break;
-      default:
-        sortByPage(_formsData, _sortBy);
-        break;
-    }
-
-    _pageAvgs = {};
-
-    for (var form in _formsData) {
-      for (var page in form.pages) {
-        _pageAvgs[page.pageName] = (_pageAvgs[page.pageName] ?? 0) + page.score;
-      }
-    }
-
-    _pageAvgs.map((pageName, score) {
-      _totalAllTeamsAvg += score;
-      return MapEntry(pageName, score / _formsData.length);
-    });
-
-    _totalAllTeamsAvg /= _pageAvgs.length;
-
-    _formsData = handleSearchQuery(_formsData, _searchQueryMap);
+    setupAverages();
 
     return DefaultTabController(
       length: 2,
@@ -106,119 +53,17 @@ class _AveragesPageState extends State<AveragesPage> {
           ),
           bottom: const TabBar(
             indicatorWeight: 4.0,
-            tabs: [Tab(text: "Averages"), Tab(text: "Compare")],
+            tabs: [Tab(text: "Insights"), Tab(text: "Compare")],
           ),
         ),
         body: TabBarView(
           children: [
             SingleChildScrollView(
-              child: Container(
-                color: GlobalColors.backgroundColor,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    const SizedBox(height: 5),
-                    DialogTextInput(
-                      onSubmit: (value) {
-                        setState(() {
-                          _searchQueryMap = evaluateSearchQuery(value);
-                        });
-                      },
-                      label: "Search",
-                    ),
-                    SingleChildScrollView(
-                      scrollDirection: Axis.horizontal,
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          DataTable(
-                            columns: [
-                              DataColumn(
-                                label: Expanded(
-                                  child: Row(
-                                    children: [
-                                      Text(
-                                        "Team",
-                                        style: TextStyle(
-                                          fontStyle: FontStyle.italic,
-                                          fontWeight: (_sortBy == "team")
-                                              ? FontWeight.bold
-                                              : FontWeight.normal,
-                                        ),
-                                      ),
-                                      IconButton(
-                                          onPressed: () {
-                                            setState(() {
-                                              _sortBy = "team";
-                                            });
-                                          },
-                                          tooltip: "Sort by Team",
-                                          icon: const Icon(Icons.sort_outlined))
-                                    ],
-                                  ),
-                                ),
-                              ),
-                              ...getPagesDataColumns(_formsData),
-                            ],
-                            rows: getDataTableRows(),
-                          ),
-                        ],
-                      ),
-                    ),
-                    const SizedBox(height: 20),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        const Text("Key (Precentile): "),
-                        const SizedBox(width: 10),
-                        Container(
-                          padding: const EdgeInsets.only(left: 5, right: 5),
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(8),
-                            color: getColorByScore(0, 100),
-                          ),
-                          child: (const Text("0-25")),
-                        ),
-                        const SizedBox(width: 10),
-                        Container(
-                          padding: const EdgeInsets.only(left: 5, right: 5),
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(8),
-                            color: getColorByScore(26, 100),
-                          ),
-                          child: (const Text("25-75")),
-                        ),
-                        const SizedBox(width: 10),
-                        Container(
-                          padding: const EdgeInsets.only(left: 5, right: 5),
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(8),
-                            color: getColorByScore(76, 100),
-                          ),
-                          child: (const Text("75-90")),
-                        ),
-                        const SizedBox(width: 10),
-                        Container(
-                          padding: const EdgeInsets.only(left: 5, right: 5),
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(8),
-                            color: getColorByScore(98, 100),
-                          ),
-                          child: (const Text("90-99")),
-                        ),
-                        const SizedBox(width: 10),
-                        Container(
-                          padding: const EdgeInsets.only(left: 5, right: 5),
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(8),
-                            color: getColorByScore(100, 100),
-                          ),
-                          child: (const Text("99-100")),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
+              child: InsightsPage(
+                allTeamsAvg: _totalAllTeamsAvg,
+                pagesAvg: _pageAvgs,
+                originalFormsData: widget.formsData ?? [],
+                calculatedFormsData: _formsData,
               ),
             ),
             SingleChildScrollView(
@@ -258,15 +103,40 @@ class _AveragesPageState extends State<AveragesPage> {
     );
   }
 
-  void sortByTeam(List<FormData> formsData) {
-    formsData.sort((form1, form2) {
-      String scouter1 = form1.scoutedTeam ?? "";
-      String scouter2 = form2.scoutedTeam ?? "";
+  void setupAverages() {
+    _totalAllTeamsAvg = 0;
+    if (widget.formsData == null) {
+      getDocuments();
+    } else {
+      _formsData = widget.formsData!;
+    }
 
-      int number1 = extractNumber(scouter1);
-      int number2 = extractNumber(scouter2);
+    _formsData = _formsData.map((match) {
+      _totalAllTeamsAvg += match.score;
 
-      return number1.compareTo(number2);
+      if (match.matchType != MatchType.normal && (match.game ?? 0) < 200) {
+        int addition =
+            MatchType.rematch == (match.game ?? MatchType.normal) ? 200 : 300;
+        match.game = (match.game ?? 0) + addition;
+        return match;
+      } else {
+        return match;
+      }
+    }).toList();
+
+    _totalAllTeamsAvg /= _formsData.length;
+
+    _formsData = calculateAvgs(_formsData).values.toList();
+    _pageAvgs = {};
+
+    for (var form in _formsData) {
+      for (var page in form.pages) {
+        _pageAvgs[page.pageName] = (_pageAvgs[page.pageName] ?? 0) + page.score;
+      }
+    }
+
+    _pageAvgs.map((pageName, score) {
+      return MapEntry(pageName, score / _formsData.length);
     });
   }
 
@@ -291,33 +161,6 @@ class _AveragesPageState extends State<AveragesPage> {
         widget.formsData = _formsData;
       });
     }
-  }
-
-  List<DataRow> getDataTableRows() {
-    List<DataRow> rows = [];
-    for (FormData form in _formsData) {
-      rows.add(DataRow(cells: [
-        DataCell(
-          TextButton(
-              child: Text(form.scoutedTeam ?? ""),
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => TeamOverviewPage(
-                      team: extractNumber(form.scoutedTeam ?? ""),
-                      forms: widget.formsData ?? [],
-                      avgs: _formsData,
-                    ),
-                  ),
-                );
-              }),
-        ),
-        ...getPagesDataRows(pages: _formsData.first.pages, data: form.pages),
-      ]));
-    }
-
-    return rows;
   }
 
   List<Widget> getAnswersWidgetsForDialog(FormData form) {
@@ -355,150 +198,6 @@ class _AveragesPageState extends State<AveragesPage> {
       ),
     );
     return answersWidgets;
-  }
-
-  void sortByGame(List<FormData> formsData) {
-    formsData.sort((form1, form2) {
-      final game1 = form1.game ?? 0;
-      final game2 = form2.game ?? 0;
-      return game1.compareTo(game2);
-    });
-  }
-
-  void sortByScouter(List<FormData> formsData) {
-    formsData.sort((form1, form2) {
-      String scouter1 = form1.scouter ?? "";
-      String scouter2 = form2.scouter ?? "";
-      return scouter1.compareTo(scouter2);
-    });
-  }
-
-  void sortByPage(List<FormData> formsData, String pageName) {
-    formsData.sort((form1, form2) {
-      double value1 =
-          form1.pages.firstWhere((page) => page.pageName == pageName).score;
-
-      double value2 =
-          form2.pages.firstWhere((page) => page.pageName == pageName).score;
-
-      return value2.compareTo(value1);
-    });
-  }
-
-  List<DataColumn> getPagesDataColumns(List<FormData> formsData) {
-    List<DataColumn> columns = [];
-
-    if (formsData.isNotEmpty) {
-      for (FormPageData page in formsData.first.pages) {
-        columns.add(
-          DataColumn(
-            label: Expanded(
-              child: Row(
-                children: [
-                  Text(
-                    "${page.pageName} Avg.",
-                    style: TextStyle(
-                      fontStyle: FontStyle.italic,
-                      fontWeight: (_sortBy == page.pageName)
-                          ? FontWeight.bold
-                          : FontWeight.normal,
-                    ),
-                  ),
-                  IconButton(
-                      onPressed: () {
-                        setState(() {
-                          _sortBy = page.pageName;
-                        });
-                      },
-                      tooltip: "Sort by ${page.pageName}",
-                      icon: const Icon(Icons.sort_outlined))
-                ],
-              ),
-            ),
-          ),
-        );
-      }
-    }
-    columns.add(
-      DataColumn(
-        label: Expanded(
-          child: Row(
-            children: [
-              Text(
-                "Total Avg.",
-                style: TextStyle(
-                    fontStyle: FontStyle.italic,
-                    fontWeight: (_sortBy == "total_score")
-                        ? FontWeight.bold
-                        : FontWeight.normal),
-              ),
-              IconButton(
-                  onPressed: () {
-                    setState(() {
-                      _sortBy = "total_score";
-                    });
-                  },
-                  tooltip: "Sort by Avg. Score",
-                  icon: const Icon(Icons.sort_outlined))
-            ],
-          ),
-        ),
-      ),
-    );
-
-    return columns;
-  }
-
-  List<DataCell> getPagesDataRows(
-      {required List<FormPageData> pages, required List<FormPageData> data}) {
-    List<DataCell> cells = [];
-    List<String> pageNames = pages.map((page) => page.pageName).toList();
-    double totalScore = 0;
-    for (FormPageData page in data) {
-      if (pageNames.contains(page.pageName)) {
-        if (data.map((page_) => page_.pageName).contains(page.pageName)) {
-          double score = page.score;
-          cells.add(
-            DataCell(
-              Container(
-                padding: const EdgeInsets.only(left: 5, right: 5),
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(8),
-                  color: getColorByScore(score, _pageAvgs[page.pageName] ?? 0),
-                ),
-                child: (Text(
-                  getNumAsFixed(score),
-                )),
-              ),
-            ),
-          );
-          totalScore += score;
-        } else {
-          cells.add(DataCell.empty);
-        }
-      }
-    }
-
-    cells.add(
-      DataCell(
-        Container(
-          padding: const EdgeInsets.only(left: 5, right: 5),
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(8),
-            color: getColorByScore(totalScore, _totalAllTeamsAvg),
-          ),
-          child: (Text(
-            getNumAsFixed(totalScore),
-          )),
-        ),
-      ),
-    );
-
-    return cells;
-  }
-
-  void sortByTotalScore(List<FormData> data) {
-    data.sort((form1, form2) => form2.score.compareTo(form1.score));
   }
 
   Map<String, FormData> calculateAvgs(List<FormData> formsData) {
@@ -557,20 +256,5 @@ class _AveragesPageState extends State<AveragesPage> {
     }
 
     return teamAvgs;
-  }
-
-  Color getColorByScore(double score, double pageAvg) {
-    double precentage = score / pageAvg;
-    if (precentage <= 0.25) {
-      return Colors.redAccent.shade400;
-    } else if (precentage <= 0.75) {
-      return Colors.transparent;
-    } else if (precentage <= 0.9) {
-      return Colors.redAccent.shade100;
-    } else if (precentage < 0.99) {
-      return Colors.greenAccent.shade400;
-    } else {
-      return Colors.blueAccent.shade100;
-    }
   }
 }
